refactor(tauri-app): extract TimelineEntry and unify the MM/DD label
- Timeline.tsx carried two self-contained components (Entry and
  OriginChip, ~150 lines) that only touch their props; they move to
  components/TimelineEntry.tsx. OriginChip stays exported because the
  view also renders it for orphaned notes.
- The 'MM/DD' slice/replace was written three times (CommandPalette's
  monthDay, Workspace's noteDate, and inline in Workspace's backlink
  list). One formatMonthDay in day-labels.ts now serves all of them.
- Two of the three 'add this date and scroll to it' sites re-spelled
  what jumpToDay already does; they call it now.

Verified in the browser harness: 84 entries, 3 origin chips and the
orphan block render as before.
